gnome: play DTMF tones on keypress
If the selected call is in progress and the
input focus is not in an entry, the client will pass
keyboard keys pressed as input to for DTMF tones
to be played by the daemon. The daemon is responsible
for filtering invalid DTMF characters.

Note: this commit disables the GTK built-in treeview
search so it does not steal input focus.

Refs #71632

Change-Id: I127ff814f8bdd84b51f1f8c62132e028189c8f68