gnome: add autostart functionality
Autostart is enabled by creating a symlink to gnome-ring.desktop
in $XDG_CONFIG_HOME/autostart (this is supported by all XDG
compliant desktop environments.
This is enabled by default and will be done the first time the
client is launched and the setting is checked on every launch.
This patch also adds a GSettings schema for the client to keep
track of the client settings (for now just the autostart enabled).

To support non-installed builds, the gschema is compiled to the
directory of the binary by default. The client first checks for
the .desktop and the compiled gschema in the local directory of
the binary.

A patch is required in the packaging branch to support the new
functionality.

Refs #74056

Change-Id: Ife9c8149e3225ab765dc1d2472ca0fd44ad35691