gnome: delete gtk rows on rowsRemoved
This patch modifies the gtk q model bindings to
delete rows in the gtk model on the rowsRemoved signal
instead of the rowsAboutToBeRemoved signal. This
seems to improve the call model view behaviour and
gets rid of the issue (in some cases) where calls which
are "over" are wrongfully selected and displayed.

Refs #75435

Change-Id: If0213f6c2112eec719063cab1e79fdf9d8e73c4a