gnome: allow client to keep running on close
Two options have been added to the client:
 - quit on close (false by default)
 - bring to foreground (true by default)

The client will now continue running by default in the background
after the user closes the main window and thus be able to receive
calls and messages. By default the client will also be brought to
the foreground when a new call is incoming.

Issue: #78461
Change-Id: I3658cd60d2df0a3e1a8c2a4f5d1173824244a3d4