gnome: reduce default window size
The window size is reduced to fit to small screens such as 800x600.

This default size is smaller than what is currently possible due to
the current layout.
It's simply there to make sure the default size is not too small
if the layout ever changes to allow it to become smaller.

Issue: #73778
Change-Id: I6d6ffbb5b419ad8548049a97a67c6467840db4b9