Update patch set 1
Patch Set 1:

This can be simplified by calling the CallModel drag and drop method with the indexes. The code for transfer/conference is already there and could be re-used. I guess it would also be possible to extend the gtkqttreemodel thing to proxy the D&D data, that way all clients could share the implementation. Note that this is not yet implemented in RecentModel (but is for contactmode, historymodel and callmodel)

Patch-set: 1
Reviewer: Gerrit User 11 <11@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=0