gnome: add delegate to handle dbus error
DBus errors usually indicate that dring has crashed or been killed.
This handler shows a warning dialog and gives dring 2.5s to restart.
If a DBus connection is not re-established, an error dialog is show and
the client quits.

This implementation prevents the client from suddenly crashing when
dring is killed due to the exception thrown by the LRC
implementation of the delegate.

Issue: #79379
Change-Id: Ie7679a11ca357d68d8b87a3925abe4598d36f02c