contacts list: restructure view and code
Place the frequent contacts, contacts, and history tree views
in separate tabs in a GtkNotebook. This simplifies the structure
of the code replacing the GtkStack and manual switching between
the views. This also removes the comboboxes which control the
sorting of the contacts and the history as they are not very
useful and are somewhat buggy.

This purpose of this patch is primarily in preparation of replacing
the calls tree view (which shows the CallModel) and the frequent
contacts view by the the "smart view", a tree view which
will show the RecentModel.

Issue: #80698
Change-Id: I76b2f6c1fcfe9fa82dba65104d928abc9df0b895