search entry: set width chars to fit placeholder text
Set the "width-chars" property to the length of the placeholder
text so that it is fully displayed, unless it is longer than 50
chars (the English version is currently 31 chars long).

This also moves the placeholder text from the .ui file into code
so that a comment can be added to the translation.

Issue: #80819
Change-Id: I0870b5505f276f16efab8d7ca1d250e29393dde6