ui: remove animations between call views
For systems running gtk+ >= 3.16 there seems to be a segfault
which happens very often, but not quite systematically at the
begining or during the animation process of the GtkStack when
it is switching between calls in progress.

Getting rid of the GtkStack widget as the container, and thus
the transition animations seems to prevent this bug. This
patch the GtkStack with a simple GtkFrame until a better fix
can be found.

Issue: #81596
Change-Id: I1bc571150fbffb3df9bf206622b31f21a8cf7c74