q tree model bindings: recursively insert rows
On rowsMoved and modelReset recursively insert all the existing children
of the top nodes. This is because rowsMoved is handled by deleting
all of the moved rows during the rowsAboutToBeMoved signal and then
re-inserting them on the rowsMoved signal. In this case, any children
of the moved rows must be re-inserted as well.

Issue: #80698
Change-Id: Ice7c8d453039acd91a55652c802fdf3e427f96e0