q tree model bindings: use rows reordered
Use gtk_tree_model_rows_reordered(), instead of deleting and then
re-inserting the moved rows, when possible to handle the rowsMoved
signal as this will keep the slection in the GtkTreeView when one
of the moved rows is selected. It also prevents us from having to
re-insert all the children recursively. However, the old technique
must be kept as this API doesn't handle the case when the parent
node of the moved rows changes.

Issue: #80698
Change-Id: I8c7924f8c25a425958adcc6d6bf1554f32dabf20