Update patch set 2
Patch Set 2:

Yes and no... Currently for each call, there is a new CurrentCallView created which inside creates a new VideoWidget for displaying the video for that call. Thus there is an instance of the VideoWidget per Call (in fact, a second VideoWidget for that Call is created if you go fullscreen... and nothing prevents you from creating more). The VideoWidget currently only knows about the 1 or 2 Video::Renderers given to it, but the Video::Renderer effectively comes from a Call.

In fact, a better solution might be to remove the handling of the "right-clicked" signal from the VideoWidget and put it into the CurrentCallView, or whichever widget contains the VideoWidget, since the CurrentCallView already contains a pointer to the Call.

Patch-set: 2