Update patch set 2
Patch Set 2:

Well, it could be, but I still think using currentModel() isn't really correct because it makes the assumption that a particular VideoWidget is necessairly tied to the selected Call, which is not the case. For example, the VideoWidget is used to show the preview in the settings, and there does not need to be an ongoing call nor a selected call to do this.

The "button-press-event" is emitted directly by the VideoWidget (not any of its internal child widgets), so it is trivial to move the handler out of the VideoWidget.

The "better" fix doesn't require any refactoring and doesn't change functionality... it simply adapts to the new API by moving the handler to a different class because that class has access to the Call object which is now needed to access the SourceModel. So its OK to have it part of this ticket, IMHO.

Patch-set: 2