main window: allow pane to be shrunk to the left and right
Most GNOME applications that use a pane allow the contents to
be shrunk so the user can resize the mane to the maximum in
either direction.

This patch also slightly increases the default window size and
allows the RingID to be ellipsized to improve the default starting
look of the main window.

Change-Id: Ib9547088640886ab956e50c3c7bf393beb0fed4c
Tuleap: #46