main window: allow pane to be shrunk left or right
Most GNOME applications that use a pane allow the contents to
be shrunk so the user can resize the mane to the maximum in
either direction.

This patch also slightly increases the default window size and
allows labels to be ellipsized to improve the default starting
layout and the behaviour when resizing the pane.

Change-Id: Ib9547088640886ab956e50c3c7bf393beb0fed4c
Tuleap: #46