ringnotify: prevent ASSERT during conferences
Do not try to connect to notifications of incoming messages
if the call is part of a conference. This actually prevents
a weird assert that happens when we try to create the
QModelIndex corresponding to the call on line 268 because the
parent is not detected as valid in the CallModel::index() function.
This patch doesn't fix the root of the issue which is in LRC.

Change-Id: Ia3202cc947d51e58302de57cfdb59274b2c32246
Tuleap: #95