video widget: make of type GtkClutterEmbed
There is no need for it be a GtkClutterEmbed in a GtkBin since
we do not use any properties of the GtkBin. It can directly be
a GtkClutterEmbed and this will also facilitate access to its
clutter scene for outside widgets. This will be used to overlay
widgets over the video in future patches.

Tuleap: #111
Change-Id: I95db46c2a6adb4a735da2980a74a5a2ca660b300