Update patch set 2
Patch Set 2:

> Should cover "Your Ring ID is:".

OK, but its not really related to this patch... I will do it in another patch.

> Getting crashes: ...

This is during the rendering of the frame... seems like a bug that was/is in the daemon/lrc (eg: the memory of the frame is already free or it is corrupted or the video renderer should have already been stopped) or else its a clutter issue. Either way it shouldn' be related to this patch... do you get the crash systematically after applying this patch?

Patch-set: 2
Label: Code-Review=0