video widget: do not listen to frameUpdated signal
The frameUpdated signal is just generated by a timer in the shmrenderer
for legacy reasons. It doesn't actually indicate that a new frame is ready.
The work of trying to fetch a new frame is done by calling currentFrame() on
the renderer which is driven by the VideoWidget timer. Thus we don't
need to listen to frameUpdated and we don't need the "dirty" atomic bool.

Change-Id: I7eff2db27151452ec270665680da9dd1a6dd8d05
Tuleap: #115