call controls: make accessible via keyboard
Allow users to navigate the call controls via the keyboard.

When the controls were moved into the clutter stage of video widget, they
lost their keyboard accessibility because clutter does not automatically
manage change of focus on tab/arrow keys nor does it manage focus chains.

This patch fixes this by handlig the "focus" signal on the video_widget and
using the focus chain of the GtkBox which contains the control buttons.

Tuleap: #119
Change-Id: I7b0b399b2a7c5a571243e653ab824678239003ee