change how fullscreen is created
The fullscreen video is now created by making the main window
fullscreen and hiding the left panel. This is done instead of
creating a 2nd GtkWindow and placing a new VideoWidget inside
of it.

This allows to easily have all the overlays and controls in
the video in the fullscreen mode.

This is also potentially a way to prevent a clutter related
crash.

Tuleap: #161
Change-Id: Iac2792b22ede4aaaff98a76d3de34189bcb4150c