Add slider to control video quality
This reverts commit 953969acc04e8ec1f9731dce770d187599e37e6d.
It also adds a checkbutton below the slider which enables
automatic video quality adjustment (and is on by default).

The slider now has a range of 0 to 100 and sets both the bitrate
and quality parameter of each codec by getting the min and max
values of both and scaling the set value.

Change-Id: I307e541c6e30c432ab5452bba2af9c2f069d79d9
Tuleap: #215