Update patch set 2
Patch Set 2:

> UserActionModel does not yet work with other models.

The UserActionModel work with all models as long as they implement your common Role API. I use it with CallModel, HistoryModel, ContactModel, CompletionModel and a few other and it work as expected. I am quite sure it work with the RecentModel too, but I don't use it for now (nothing against it, I just have some proxies on top of the other ones that provide most of the features it has).

Patch-set: 2
Label: Code-Review=0