Update patch set 3
Patch Set 3:

OK, I'll give it a try... Its just that I saw that in the constructor of the UAM, it connected to some signals of the CallModel and then there was a TODO for the other models. Anyways, I'll give it a shot in another patch.

Patch-set: 3
Label: Code-Review=0