out of call chat
Initial implementation. Now clicking on an item in the Conversations
view (RecentModel) will bring up a chat view. In the case of a Person
(contact) it will select the chat with the last used ContactMethod of
that person. If there is more than one ContactMethod, as combo box will
be displayed giving the choice of ContactMethods to use.

To make a call, double-click the item as before. Any call (incoming or
outgoing) will superseed the chat view.

Out of call chats use the account based chat API. In call chats still
use the call based chat API.

Change-Id: I3deb09fd22c3dda7b78ea9be0eef32a6f27adecb
Tuleap: #203