mainwindow: improve view updates
Connect to the dataChanged signal of the RecentModel instead of the
callStateChanged and incomingCall signals of the CallModel. The
RecentModel is driven by the signals of the CallModel, so in certain
cases we could receive those signals before the RecentModel and the
RecentModel will not yet be updated properly, causing an incosistent
view. Since the view is now driven by the RecentModel, we should
connect to its signals instead of that of the CallModel.

Change-Id: I3db7152a7429623ba95822597160647f8be65487
Tuleap: #203