chat: show unread message count in contact image
Renders a notification showing the number of unread chat messages
in the bottom right corner of the contact's image.

Messages which are printed to the chat view are automatically marked
as read.

The chat view now also only attempts to print text messages, rather
than messages of all mime types.

Tuleap: #203
Change-Id: I0592257c0f5bff9deb3110aec4ba483fddc95e02