allow main window to be resized smaller
Certain views/widgets were preventing the ability to resize the
main window to a smaller size. They have either been placed in
resizable containers (GtkScrolledWindow or GtkPaned) or their
properties have been modified such that they can be shrunk.

The account view has also been made to look more consistent with
the rest of the application as a result.

Change-Id: Id8ee182becb1bee58a9fbf66236d795df54fc0f5
Tuleap: #348