account wizard: remove second step
The ring ID is shown and explained in the welcome view of the main
window now, so there is no need for a second step in the account
wizard process to explain it.

This also allows users to leave their username blanc and in this
case it is set to 'Unknown'.

Change-Id: Id25b49b2663e81990847dddb89cb1997d449b155
Tuleap: #366