Update patch set 9
Patch Set 9: Code-Review-1

libqrencode is packaged in Ubuntu, and so I assume all our supported distros (to be verified). If the minimum version that we need is available in all the distros we support, we should use the packaged version instead of including it as a submodule, as this would have to be done for official packages anyways.

Patch-set: 9
Label: Code-Review=-1