Update patch set 14
Patch Set 14: Code-Review-1

(2 comments)

So, the bootstrap correctly shows up for an existing account, but its empty if I create a new account. Its likely an init problem in the bootstrap model / new account creation in LRC, but it should probably be fixed before we merge this patch.

Patch-set: 14
Label: Code-Review=-1