Add Ring systray icon
This adds an icon in the notification areas (systray) by default.
The icon also has a popup menu which allows the user to show or hide
the main widow and to quit Ring.

It uses 2 systray icon implementations:
- AppIndicator from Canonical
- GtkStatusIcon from Gtk, but deprecated

A CMakeLists option is added "USE_APPINDICATOR" and is ON by default.
However this adds a dependency on appindicator3-0.1. If the option is
set to OFF or the dependency is not found then we fallback on using
GtkStatusIcon.

AppIndicator is prefered so that the status icon appears for Unity
users, as unity does not support GtkStatusIcon by default.

Additionaly, the GSettings schema has been modified slightly. The
general Ring settings menu also now says has the option "Show Ring
icon in the notification area (systray)" instead of of "Hide Ring
on close instead of quitting". This mimics other gtk app behavior,
such as Transmission. When the systray is enabled, closing the main
window will not quit Ring. When the systray is disabled closing the
main window will quit Ring so that it doesn't continue to run
withou the user being aware.

Tuleap: #370
Change-Id: I9f5c2811e135d20b0b6c81ab4a80be748f6a1be3