Update patch set 9
Patch Set 9:

ok, I fixed the minor issue with the menu object being dereferrenced.

Otherwise, I can confirm there is an issue with the menu of the icon on Unity, however, it seems to be an issue with Unity not updating the menu if the action is not done via the menu. On GNOME 3 it works properly.

Patch-set: 9
Reviewer: Gerrit User 23 <23@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=0