search entry: open chat view instead of call
This changes the default behaviour of the search entry, the button
next to it, and the autocompletion drop down to opening the chat view
with the selected CM/Person instead of placing a new call. This allows
people who wish to simply chat with a new RingID to do so, instead of
first having to call that RingID or add it to a Contact in their
addressbook.

Additionally, a setting has been added to change this behaviour, so
that (typically) SIP users can easily place new calls by entering phone
numbers in the search entry.

Also, a "Place call" button has been added to the chat view (when not
in a call), so that users can easily call someone they are chatting with
without having to double click on that contact in the contacts list.

Change-Id: Ia833fb36620fd34afdbb3c3a4357c212a87f8796
Tuleap: #953