monitor network status
Monitor the network status and notify the daemon when it changes. To
prevent from notifying the daemon too often, we wait until there are
no network changes for a timeout of 3 seconds. Also, since its
possible that we still have connectivity (eg: if we connect to a new
network but we're still also connected to the previous one), we wait
until there are no more calls in progress.

Change-Id: Id0f222a90cda2822f2f4186ad7da7265e485d075
Tuleap: #845