monitor network status using libnm
We monitor the primary connected network and notify the daemon when
it changes so that it can re-establish account connections. We first
check there arent any calls in progress to prevent the daemon
terminating calls which are still connected.

This functionality is achieved using libnm, thus libnm is added as
an optional dependency.

Change-Id: I60233801c13740619b2e6473f5bc6d17901d7bf3
Tuleap: #845