avatar: fix black camera image
On some systems, when opening the camera to take a new picture for
the avatar, the result is a black image. The log also outputs
"bad dawable". It seems the VideoWidget doesn't like being a page
directly in a GtkStack. Instead we put it in a GtkFrame which does
not get destroyed. This seems to fix the issue.

Change-Id: Ia3796ad3b3f02947619c2b0ba57ad4b54e30552e
Tuleap: #840