prevent settings closing after first launch
This fixes the weird settings bug happening when Ring was being used
for the first time and the account settings would close any time the
account settings were modified. The problem was simply that the
lambda to the Account::changed signal was not disconnected after
it was no longer needed.

Change-Id: Ie37dcd52819a71920c6c73032e3f0b11caa362f9
Tuleap: #935