improve context menu in treeview behaviour
Now rightclicking on an item in the contact treeviews (Conversations,
Contacts, History) will result in the item clicked on being selected.
Before the popup menu would open for the selected item, but the
selection would not change if the click was on a different item.

This also fixes a small memory leak where the a new menu widget would
be created on every right click, but the old menu was never destroyed.
Now the same menu widget is used, it is simply updated once the
selection is changed.

Change-Id: I4e54d618a090e28b565cbef719065c943a826b0e
Tuleap: #930