Update patch set 26
Patch Set 26:

One issue I am aware of is that when you try to import an existing account and enter a bad pin or password, it will create an invalid account.

The account is actually deleted daemon-side, but the change is not reflected in the client.

I think this is unrelated to this issue and should be fixed separately. The client should react to accounts being deleted by the daemon.

Patch-set: 26
Label: Code-Review=0