show chat/call view from selection in history and contacts
Previoiusly, the chat/call view displayed could only be changed by
changing the selection in the converstaions list view. The contacts
and history list views would only change the call view if a new
call was initiated by double clicking on the selection.

This patch changes this behaviour, now slecting an item in any of
the three contact list views will result in showing the chat view
with that Person or ContactMethod, or showing the call view if there
is an ongoing call with the selected item.

In order to prevent conflicts between different selections in the
different views, only one selection at a time is allowed. So,
changing the selection in the conversations view, for example, will
clear the selection in the other 2 list views except in the case
that the item selected is the same one. For example, if the Person
selected in the Contacts view is the same as the one in the
Conversations view, then both selections will remain.

Change-Id: Icc00c5801e66ca0013b730d6d008ebf904a999e1
Tuleap: #965