use CcCropArea to crop avatar
CcCropArea code is taken from gnome-control-center code.

It is used to crop avatar images in the control center and in
gnome contacts.

Using it unifies the interface more with other gnome programs and
also gets rid some of the small bugs in the previous implementation.

Now instead of creating a new selection each time, the user simply
moves and resizes one selection.

Change-Id: I764e958cf9e5e6f1aadd754ddd1ad5d542415365
Tuleap: #917