quality: remove QtkQSortFilterTreeModel
This Qt model wrapper class was added because GtkQTreeModel didn't
seem to work for QSortFilterProxy models; however this does not seem
to be the case any longer. Either something was fixed in the
GtkQTreeModel code, or else the initial problem was due to another
bug. Either way this reduces and simplifies the code base.

Change-Id: I8d1f7bb83c17708e2ef3c544ee32c13a4b2b3e7d