Update patch set 10
Patch Set 10: Code-Review-1

(1 comment)

The text wrapping mode seems to be set to by word... so if the txt msg contains a really long word, it will result in having to scroll the view horizontally... is it possible to set the wrapping to character, so that doesn't happen? You could set a minimum size to the scroll view so that the view can't be made so small that everysing char is a new line.

Patch-set: 10
Label: Code-Review=-1