Update patch set 1
Patch Set 1:

> Why does this need to be done?

The daemon has changed the account username and we need to take it into account.

Alright so Reloading the account every time the state has changed is dangerous because we could do it while its being edited.

So the right thing to do would be to update the welcome view after the account creation reload. (this one we have to keep).

Patch-set: 1
Label: Code-Review=0