Update patch set 2
Patch Set 2: Code-Review-1

I feel like its weird to be driving an interface which shows the data of a model/object by another interface rather than by that model/object. It seems the issue is simply that the 3 signals used currently (accountListUpdated, accountEnabledChanged, registrationChanged) to update the view are not enough, since its possible that the ringID gets set after they are called. I used those 3 signals initially so that I wouldn't have to use the "dataChanged" signal and thus update the view in cases the active account didn't actually change.

Thus it seems the solution is to:
a. simply connect to "dataChanged" of the AccountModel
b. keep track of the active account and connect to its "changed" signal
c. add some kind of "accountCreationComplete" signal to indicate when the ringID has been generated (this could also be used to simplify the current migration/creation code which uses the account state to figure this out)

Patch-set: 2
Label: Code-Review=-1