new release display and information
This patch changes the way to display the release information.
Also update the release name to Louis Riel - beta2

Now the revision is replaced by the build date.

This value is generated as following pseudo-code:

if SOURCE_DATE_EPOCH is set
  it's value must be a number of seconds since epoch.
  it's used to generate the final build date displayed.
else if RING_CLIENT_BUILD_DATE is set
  it's value is used as-it as the final build date displayed.
else
  the current date of the system is used to generate the final build date displayed.

Change-Id: Ifa431b8fa4f8945cf9a25fef028712ad0e40a0f3