Display Ring username after account creation
The Account settings view was being created during account creation
but before registration and the username box was not being updated
once the registration completed. We now connect to the signal
earlier and update the username box accordingly once registration is
complete, even if registration was not initiated from the username
box widget.

Change-Id: I619cbf065d2b2eb6df8399cac0b537d8219eabca
Tuleap: #1177