Specify Ring account when perform name lookup
Use the default Ring account (normally the first Ring account which
is enabled and registered) when performing the name lookup in the
search entry. This way the name server of that account will be used
instea of the default name server. Also this is the account which
should normally be used to send the message / make the call.

Reviewed-by: Alexandre Viau <alexandre.viau@savoirfairelinux.com>
Change-Id: I40b7fc893803921293ff02128ef1a709e7b914db