chatview: prevent crash due to invalid memory
I'm not sure about this one, but the documentation of QString.constData()
says that the returned pointer is only valid as long as the QString object
is not modified:
http://doc.qt.io/qt-5/qstring.html#constData

In this case its possibly that maybe the compiler is optimizing somewhere
and the QString object is destroyed before the constData pointer is used.

I noticed that in most of the rest of the code, the constData pointer is
used immediately, instead of first being stored somewhere.

Change-Id: I678f33c91d3be61814d405d3b465af485937814d
Tuleap: #1349