AccountView: removed a gtk_widget_show_all call
This removes one call to gtk_widget_show_all in AccountView. It
prevented disabling the visiblity of some widgets in the tabs in their
constructors because gtk_widget_show_all would set the visibility to
true again.

While this could be done with a widget property (I forgot which one), I
I think its cleaner to get rid of the gtk_widget_show_all calls when
possible, in order to avoid unexpected behaviours as much as possible.

Change-Id: Ida7bf85dfc322b672732ee4f70421dec6b0bb0a7