Account Settings: Remove Audio and Video tabs
Remove Audio and Video tabs and move their functionality to the
advanced tab. These features should almost never be touched and they
don't "deserve" their own tab, as if they were important.

I think that this is a much better compromise than to remove the feature
altogether, because this could always be useful to power users.

Set the bootstrap servers treeview to have a matching 10px margin

Tuleap: #1408
Change-Id: Ia247f474b050c9d8494b8289055cc8b9a619cd20