Only use AppIndicator in Unity
Using AppIndicator in other DE (at least in GNOME) seems to cause
a bizzare critical log message at launch. To prevent this we use
the GtkStatusIcon API instead unless we're on Unity.

Note that this change removes the 'USE_APPINDICATOR' CMake variable.
We will always check for the presence of the lib, but only use it
on Unity.

Change-Id: Icb2e56853a5f2ac16d74fdd1b9d5c5d17a02479e
Tuleap: #1440