Update patch set 2
Patch Set 2:

(5 comments)

I don't seem to get any sort of indicator that account migration is in process. When I click "migrate" the daemon starts migrating the account, but I can keep putting in other passwords and click migrate again. Also, when migration completes, nothing happens in the UI until the timeout:
** (gnome-ring:19614): WARNING **: timeout reached while migrating account 8bea033e688567d8

Not sure if this is an issue with the daemon again, but the Ring-id feild is always empty on the migration screen.

This is a minor point and is more of a design issue, but the migration screen now just says "Please enter your password:". But this implies that the user already has a password... Its not clear what this password is. I think it should be clearly indicated that the user is creating a new password for this account.

Patch-set: 2