Update patch set 2
Patch Set 2:

(1 comment)

okay, so I retested with the correct type of beta2 account

please still see my comments in the .cpp file

Also, when the migration fails (wrong password), the dialog about the wrong password is shown for just a split second, and then the migration screen comes up again. Either this dialog needs to be shown until "OK" is pressed. Although I think the easier way visually would be to just display an error text bellow the password entry saying the password is wrong. Either way, this should be fixed.

Patch-set: 2