config.h: fix check for appindicator
If appindicator lib is not found APPINDICATOR_FOUND variable is not
set by cmake. Thus HAVE_APPINDICATOR in config.h is not set. This
seems to default to 0 in the pre-processor, and thus to true. So in
this case we try to build with appindicator, even though it doesn't
exist.

We fix this by explicitly setting HAVE_APPINDICATOR to 0 or 1.

Change-Id: I50baf0262f95f899262c71f70783dcccb11c53ee