Update patch set 2
Patch Set 2: Code-Review-1

this needs to be rebased on the other patch (there is a conflict)

but I'm also wondering if we want to get rid of this timeout? what if the daemon gets stuck for some other reason? eg: I think it gets stuck if you try to migrate a beta1 account. I know this isn't supported anymore... but it kinda sucks if it gets stuck on the migrate account screen and its impossible to do anything... at least if it times out the user can then go and delete their old beta1 account and create a new account

Patch-set: 2
Label: Code-Review=-1