Update patch set 5
Patch Set 5: Code-Review-1

(1 comment)

looks good, aside from minor nitpick

IMHO, its a bug in the daemon that the query never ends when there is no connection to the name server. This bug has been there since we introduced the name server in the daemon and I've brought it up before. The daemon should time out and return an error like it does when there is no matching name found.

Patch-set: 5
Label: Code-Review=-1