fix search bar crash
Private structure of main window has a direct C++ object.
But GTK is a C world, so this object is zero initialized,
but not constructed using C++ constructor.
Unfortunatly the type is str::string and some implementation
like the GNU libstdc++ are not well defined if zero-initialized.

This causes typicaly a crash in the main window code when
search bar is used with a RingID (ie SHA1 hash).

The patch fixes that by using a pointer correcly allocated
with C++ new operator.

Change-Id: I40e9603b0ab3d04c19dca1843163f286c5266afd