Update patch set 23
Patch Set 23:

(4 comments)

> (1 comment)

This is not a reason for a -2.
We are in business pressure to release, focus on bugs first.
Code cleanup, naming convention and related stuff are importants, but not blocking (-2). -1 is correct to warn about the fact and obtain a technical debt.

Patch-set: 23
Reviewer: Gerrit User 3 <3@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=0