Update patch set 23
Patch Set 23:

Once used by clients, the API is Hell to change. This name thing was clearly stated from the beginning, and frequently repeated. I also personally warned about this specific issue two weeks ago and was told the name would be changed soon.
This is not a variable or method name issue. It's an important API issue.

Patch-set: 23