Update patch set 9
Patch Set 9:

This is not a variable or method name issue. It's an blocking API issue. This is not the place to warn : I already warned about this "Contact request" naming issue many times, including about this specific LRC API issue. Now this needs to be fixed before we can start merging client patches using this API.

Patch-set: 9