prevent crash on Wayland on call start
There is a strange crash on Wayland during the gtk main loop when
a call starts. Its not clear what the cause is, but combining the
following two changes seems to get rid of it:

- revert from using the same instance of WebKitWebView to a new
  instance for each ChatView

- remove all webkit animations in chat view

This seems to fix the crash when a call starts when running on
Wayland.

Change-Id: Iaea6a39ff1e4670d7a0848b9008a51470b259e16
Tuleap: #1320