Update patch set 13
Patch Set 13: Code-Review-1

I do not think it is up to the client to choose the selected account but to the user. If a call arrives for another account, it must be able to respond with an active account without changing the selected account.

Patch-set: 13
Reviewer: Gerrit User 126 <126@1e26b4c1-6b8c-47ff-9df7-501b633d3b5a>
Label: Code-Review=-1