Update patch set 5
Patch Set 5: Code-Review-1

some issues with auto scroll:

* when a conversation is selected, the chat view scrolls from beginning to end with a weird animation. it get worst as the conversation grows.
* when a conversation is smaller than the window it aligns at the top instead of bottom;
* when the whole conversation fits in the windows and the window is
resized to a smaller size, the top of the conversation is shown instead of the bottom. If the conversation is scrolled to the bottom (default view) and the window is resized to a smaller size, the conversation scrolls up.

Patch-set: 5
Label: Code-Review=-1